clean(ZMSKVR): Refactor some components to a better place in zmscitizenview
