+++ zmscitizenview/tests/unit/Appointment/ServiceFinder/AltchaCaptcha.spec.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, vi } from "vitest";
 import { nextTick } from "vue";
 
 // @ts-expect-error: Vue SFC import for test
-import AltchaCaptcha from "@/components/Appointment/AltchaCaptcha.vue";
+import AltchaCaptcha from "@/components/Appointment/ServiceFinder/AltchaCaptcha.vue";
 
 // Mock window.scrollTo for jsdom
 globalThis.scrollTo = vi.fn();
